# openapi.yaml
...
openapi: 3.1.0
info:
  title: CogniGate API
  version: 1.0.0
  summary: 'Self-hosted OpenAI-compatible gateway: routing, quotas, caching and metering across
    providers.'
  description: |
    One process serves both planes on one port.

    - **The data plane**, under `/v1`, is OpenAI-compatible. An existing SDK pointed at this base URL works unchanged, errors included.
    - **The admin plane**, under `/admin/v1`, configures tenants, keys, providers, aliases, routes, quotas and webhooks.

    A key is minted for exactly one plane and is never valid on the other; presenting a data key to `/admin/v1` fails with `wrong_plane` rather than with a 404.

    This document describes every route the gateway serves. A reconciliation test compares it against the running route table in both directions, so a route added without an entry here — or an entry here for a route that does not exist — fails the build.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: http://localhost:8080
  description: The reference deployment. Both planes are on this one origin.
...
paths:
...
  /v1/usage:
    get:
      tags:
      - Usage
      operationId: getUsage
      summary: Usage and quota state for the calling tenant.
      description: |
        The same window the quota rejections are computed over, so a caller asking why it was refused sees the figures the refusal used.
      security:
      - DataKey: []
      parameters:
      - name: window
        in: query
        description: The period to report over.
        schema:
          type: string
          enum:
          - day
          - month
          default: day
      responses:
        '200':
          description: Totals and every quota slot that applies.
          headers:
            X-CogniGate-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/usage/breakdown:
    get:
      tags:
      - Usage
      operationId: getUsageBreakdown
      summary: Usage broken down by one dimension.
      description: |
        Rows are ordered by spend. At most 200 are returned; when there were more, `truncated` is true and the cheapest were dropped.
      security:
      - DataKey: []
      parameters:
      - name: window
        in: query
        description: The period to report over.
        schema:
          type: string
          enum:
          - day
          - month
          default: day
      - name: group_by
        in: query
        description: The dimension to aggregate on.
        schema:
          type: string
          enum:
          - model
          - provider
          - key
          - client_request_id
          default: model
      responses:
        '200':
          description: The breakdown.
          headers:
            X-CogniGate-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageBreakdown'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /admin/v1/tenants/{tenant}/usage/breakdown:
    get:
      tags:
      - Usage
      operationId: getTenantUsageBreakdown
      summary: Usage for one tenant, broken down by one dimension.
      security:
      - AdminKey: []
      parameters:
      - name: tenant
        in: path
        required: true
        description: Tenant id. A root key reaches any tenant; a `tenant:<id>` key reaches
          only its own, and naming another is a 404.
        schema:
          type: string
      - name: window
        in: query
        description: The period to report over.
        schema:
          type: string
          enum:
          - day
          - month
          default: day
      - name: group_by
        in: query
        description: The dimension to aggregate on.
        schema:
          type: string
          enum:
          - model
          - provider
          - key
          - client_request_id
          default: model
      responses:
        '200':
          description: The breakdown.
          headers:
            X-CogniGate-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageBreakdown'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
components:
  securitySchemes:
    DataKey:
      type: http
      scheme: bearer
      description: |
        A data-plane key, prefixed `cg-`. Valid on `/v1/*` only. The prefix is part of the credential rather than decoration: it is how the gateway answers "wrong plane" without a store lookup, and how secret scanners recognise a leaked CogniGate key in a public repository.
    AdminKey:
      type: http
      scheme: bearer
      description: |
        An admin-plane key, prefixed `cga-`, or the deployment's bootstrap key. Valid on `/admin/v1/*` only. Scope is either `root`, which reaches every tenant, or `tenant:<id>`, which reaches one.
  headers:
    RequestId:
      description: |
        The identifier that ties this response to its structured log line, its usage record and, on a failure, the `request_id` in the error body. It is the one string a user can quote about a request.
      schema:
        type: string
...
    QuotaState:
      description: |
        Where the caller stands against its GW-4 quotas after this request. `soft-exceeded` is the warning that arrives while requests are still being served.
      schema:
        type: string
        enum:
        - ok
        - soft-exceeded
        - hard-exceeded
...
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
  responses:
    BadRequest:
      description: The request is malformed, or a parameter is outside the range the deployment
        accepts.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No credential, an unknown one, or one minted for the other plane (`wrong_plane`).
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential is valid but its scope does not reach this resource.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No such resource. A tenant-scoped key naming another tenant also gets this,
        rather than a 403, so the response does not reveal whether that tenant exists.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: Rate limited, at the concurrency ceiling, or over a hard quota. `code`
        says which.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-CogniGate-Quota-State:
          $ref: '#/components/headers/QuotaState'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ServiceUnavailable:
      description: The process is draining, or a dependency it needs is unreachable.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - message
          - type
          - code
          - param
          properties:
            message:
              type: string
              description: Human-readable explanation. Never contains prompt or completion
                content.
            type:
              type: string
              description: OpenAI's error family, reused so an existing SDK branches on it
                unchanged.
              enum:
              - invalid_request_error
              - authentication_error
              - rate_limit_error
              - api_error
            code:
              type: string
              description: The closed CogniGate code registry. A client that branches on anything
                should branch on this.
              enum:
              - invalid_api_key
              - wrong_plane
              - invalid_request
              - fallback_duplicate_model
              - capture_ttl_too_long
              - unsupported_parameter
              - insufficient_scope
              - model_not_found
              - alias_unresolvable
              - not_supported
              - resource_not_found
              - alias_collides_with_model
              - conflict
              - request_too_large
              - rate_limited
              - concurrency_exceeded
              - quota_exceeded
              - budget_exceeded
              - upstream_exhausted
              - response_too_large
              - upstream_error
              - service_unavailable
              - gateway_timeout
              - upstream_stream_stalled
            param:
              type:
              - string
              - 'null'
              description: The offending field, or null when the failure is not attributable
                to one. Null rather than omitted, because OpenAI clients expect the key to
                be present.
            request_id:
              type: string
              description: Echoes X-CogniGate-Request-Id.
            attempts:
              type: array
              description: 'Present on `upstream_exhausted` only (GW-3.AC-5): the cascade
                as it was tried, in order.'
              items:
                $ref: '#/components/schemas/Attempt'
      description: |
        The GW-7 error envelope. Every failure on either plane leaves the gateway in this shape, so stock OpenAI error handling parses a CogniGate error without a special case.
    Attempt:
      type: object
      required:
      - provider
      - model
      - failure
      properties:
        provider:
          type: string
        model:
          type: string
        failure:
          type: string
          description: Classification of why this candidate failed, from a closed vocabulary.
            Never the upstream's own message.
        status:
          type: integer
          description: The upstream's HTTP status; absent when there was never a response
            to take one from.
...
    UsageLimit:
      type: object
      required:
      - scope
      - window
      - unit
      - cap
      - soft_threshold_pct
      - consumed
      - remaining
      - resets_at
      - state
      properties:
        scope:
          type: string
          enum:
          - tenant
          - key
        window:
          type: string
          enum:
          - day
          - month
        unit:
          type: string
          enum:
          - tokens
          - cost
        cap:
          type: number
        soft_threshold_pct:
          type: integer
        consumed:
          type: number
          description: Consumption in this slot's own unit. A cost slot follows `billing.cost_visibility`,
            so it can be a rounded figure on the data plane.
        remaining:
          type: number
          description: '`cap` minus `consumed`, floored at zero, so the two always sum to
            the cap as reported.'
        resets_at:
          type: string
          format: date-time
        state:
          type: string
          enum:
          - ok
          - soft-exceeded
          - hard-exceeded
    UsageResponse:
      type: object
      required:
      - object
      - window
      - since
      - until
      - state
      - limits
      - requests
      - prompt_tokens
      - completion_tokens
      - total_tokens
      - cost_usd
      - charge_usd
      properties:
        object:
          type: string
          const: usage
        window:
          type: string
          enum:
          - day
          - month
        since:
          type: string
          format: date-time
        until:
          type: string
          format: date-time
        requests:
          type: integer
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
        cost_usd:
          type: number
          description: 'What the traffic cost at the provider''s published rate. Zero for
            a model whose listing publishes no price and whose rate the operator has not supplied.
            This is the figure cost quotas are measured against. A deployment that sets `billing.cost_visibility:
            hint` rounds it to one significant figure on the data plane; the admin plane and
            the stored row are always exact.'
        charge_usd:
          type: number
          description: What is owed for the traffic. Equal to `cost_usd` unless the operator
            has put a margin on the provider rate, or is carrying the bill themselves, in
            which case it is zero.
        state:
          type: string
          description: The worst state across `limits`.
          enum:
          - ok
          - soft-exceeded
          - hard-exceeded
        limits:
          type: array
          description: Every quota slot that applies, tenant and key alike. Empty when nothing
            is capped.
          items:
            $ref: '#/components/schemas/UsageLimit'
...
    UsageBucket:
      type: object
      required:
      - key
      - requests
      - prompt_tokens
      - completion_tokens
      - total_tokens
      - cost_usd
      - charge_usd
      properties:
        key:
          type: string
          description: The value of the grouping dimension for this row.
        requests:
          type: integer
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
        cost_usd:
          type: number
          description: 'What the traffic cost at the provider''s published rate. Zero for
            a model whose listing publishes no price and whose rate the operator has not supplied.
            This is the figure cost quotas are measured against. A deployment that sets `billing.cost_visibility:
            hint` rounds it to one significant figure on the data plane; the admin plane and
            the stored row are always exact.'
        charge_usd:
          type: number
          description: What is owed for the traffic. Equal to `cost_usd` unless the operator
            has put a margin on the provider rate, or is carrying the bill themselves, in
            which case it is zero.
    UsageBreakdown:
      type: object
      required:
      - object
      - window
      - group_by
      - since
      - until
      - data
      - truncated
      properties:
        object:
          type: string
          const: usage_breakdown
        window:
          type: string
          enum:
          - day
          - month
        group_by:
          type: string
          enum:
          - model
          - provider
          - key
          - client_request_id
        since:
          type: string
          format: date-time
        until:
          type: string
          format: date-time
        data:
          type: array
          description: Ordered by spend, descending.
          items:
            $ref: '#/components/schemas/UsageBucket'
        truncated:
          type: boolean
          description: True when there were more than 200 buckets and the cheapest were dropped.
...